# src/main/resources/org/damap/base/dmpcommonstandard/openapi.yaml
openapi: 3.1.4
info:
  title: Common Machine-Actionable DMP API
  description: |
    This API provides access to Data Management Plans in a machine-readable format. You can use this API to perform
    automated analysis or other integration. Please note that this API may not give you all information stored in the
    system implementing it as this standard is intended as a common baseline among multiple DMP platform vendors.

    # Schema extension

    This API supports extending the standard by using content negotiation. The MIME types used for content negotiation
    *should* follow the vendor tree naming defined in
    [RFC 6838 Section 3.2](https://datatracker.ietf.org/doc/html/rfc6838#section-3.2) and make use of the media type
    suffixes defined in [RFC6839](https://www.rfc-editor.org/rfc/rfc6839). The vendor prefix should be chosen such that
    the vendor is clearly identifiable. For example:

    ```
    application/vnd.com.example.dmp+json
    ```

    Clients *should* send an `Accept` header with their preferred MIME types. The server *must* order the contents of
    the `Accept` header correctly and select the standard with the highest client preference.
    (See the [MDN docs](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Accept) for a detailed
    explanation on this header.)

    If the client sends no `Accept` header or sends an unqualified `application/json` request, the server *may* select
...

...

    ```
...
    ```

...

...

...

...
  version: 0.1.0
...
security:
  - BearerAuth: []
paths:
  /dmps:
    get:
      summary: List/search DMPs
      description: |
        This endpoint lists all DMPs or allows for creating a filtered list. When filters are provided,
        all filter are applied (AND relationship). For filters supporting lists, the individual values are applied
        as an OR relationship.

        For items accepting more than one value you may pass multiple values by repeating the parameter in the
        query string for each item.
      operationId: listDMPs
      tags:
        - DMP
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip from the start.
          schema:
            title: Offset
            type: integer
            minimum: 0
            default: 0
        - name: count
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items to fetch.
        - name: sort[]
          description: |
            A list of fields to sort by with the sort order.
          in: query
          required: false
          schema:
            type: array
            default: ["created,desc"]
            items:
              type: string
              default: created,desc
              enum:
                - title,asc
                - title,desc
                - created,asc
                - created,desc
                - modified,asc
                - modified,desc
                - language,asc
                - language,desc
                - embargo,asc
                - embargo,desc
                - keyword,asc
                - keyword,desc
        - name: created_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: created_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: modified_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: modified_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: languages[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/LanguageCode"
        - name: contact_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ContactID"
        - name: contributor_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/ContributorID"
        - name: dataset_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/DatasetID"
        - name: metadata_standard_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/MetadataStandardID"
        - name: dmp_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/DMPID"
        - name: funder_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/FunderID"
        - name: grant_ids[]
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/GrantID"
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: ethical_issues_exist
          in: query
          required: false
          schema:
            type: boolean
        - name: embargo_before
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: embargo_after
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        200:
          $ref: "#/components/responses/DMPListResponse"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/AuthenticationRequiredResponse"
        406:
          $ref: "#/components/responses/NotAcceptableResponse"
        default:
          $ref: "#/components/responses/GenericErrorResponse"
    post:
      summary: Create a DMP
      description: |
        Create or import a DMP from a JSON-formed DMP object. The implementation may opt to allow the creation of
        draft DMPs with formally valid, but incomplete fields. Alternatively, the implementation may reject incomplete
        DMPs if it cannot store them.
      operationId: createDMP
      tags:
        - DMP
      requestBody:
        required: true
        content:
          application/vnd.org.rd-alliance.dmp-common.v1.2+json:
            schema:
              $ref: "#/components/schemas/DMPDocument"
          application/json:
            schema:
              $ref: "#/components/schemas/DMPDocument"
      responses:
        200:
          description: DMP successfully created
          headers:
            Last-Modified:
              description: |
                The server should expose the time the DMP was last modified to allow for a conditional PUT request.
                This is especially important if the server does not support PATCH requests.
              schema:
                type: string
                format: date-time
          content:
            application/vnd.org.rd-alliance.dmp-common.v1.2+json:
              schema:
                $ref: "#/components/schemas/DMPWithID"
            application/json:
              schema:
                $ref: "#/components/schemas/DMPWithID"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/AuthenticationRequiredResponse"
        403:
          $ref: "#/components/responses/InsufficientPermissionsResponse"
        406:
          $ref: "#/components/responses/NotAcceptableResponse"
        409:
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/GenericErrorResponse"
  /dmps/{id}:
    get:
      summary: Get a DMP
      description: |
        Fetch a DMP based on its ID.
      operationId: getDMP
      tags:
        - DMP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/DMPSystemID"
      responses:
        200:
          description: DMP retrieved
          headers:
            Last-Modified:
              description: |
                The server should expose the time the DMP was last modified to allow for a conditional PUT request.
              schema:
                type: string
                format: date-time
          content:
            application/vnd.org.rd-alliance.dmp-common.v1.2+json:
              schema:
                $ref: "#/components/schemas/DMPWithID"
            application/json:
              schema:
                $ref: "#/components/schemas/DMPWithID"
        401:
          $ref: "#/components/responses/AuthenticationRequiredResponse"
        403:
          $ref: "#/components/responses/InsufficientPermissionsResponse"
        404:
          $ref: "#/components/responses/DMPNotFoundResponse"
        406:
          $ref: "#/components/responses/NotAcceptableResponse"
        default:
          $ref: "#/components/responses/GenericErrorResponse"
    put:
      operationId: putDMP
      tags:
        - DMP
      summary: Overwrite a DMP
      description: |
        Completely overwrite a DMP with the specified data. The server MUST honor the If-Unmodified-Since and IF-Match
        headers if present.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/DMPSystemID"
        - name: If-Unmodified-Since
          in: header
          required: false
          schema:
            type: string
            format: date-time
      requestBody:
        required: true
        content:
          application/vnd.org.rd-alliance.dmp-common.v1.2+json:
            schema:
              $ref: "#/components/schemas/DMPDocument"
          application/json:
            schema:
              $ref: "#/components/schemas/DMPDocument"
      responses:
        200:
          description: DMP modified
          headers:
            Last-Modified:
              description: |
                The server should expose the time the DMP was last modified to allow for a conditional PUT request.
                This is especially important if the server does not support PATCH requests.
              schema:
                type: string
                format: date-time
          content:
            application/vnd.org.rd-alliance.dmp-common.v1.2+json:
              schema:
                $ref: "#/components/schemas/DMPWithID"
            application/json:
              schema:
                $ref: "#/components/schemas/DMPWithID"
        400:
          $ref: "#/components/responses/BadRequestResponse"
        401:
          $ref: "#/components/responses/AuthenticationRequiredResponse"
        403:
          $ref: "#/components/responses/InsufficientPermissionsResponse"
        404:
          $ref: "#/components/responses/DMPNotFoundResponse"
        406:
          $ref: "#/components/responses/NotAcceptableResponse"
        409:
          $ref: "#/components/responses/ConflictResponse"
        415:
          $ref: "#/components/responses/UnsupportedMediaTypeResponse"
        default:
          $ref: "#/components/responses/GenericErrorResponse"
    delete:
      operationId: deleteDMP
      tags:
        - DMP
      summary: Delete a DMP
      description: Delete a DMP based on its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/DMPSystemID"
      responses:
        204:
          description: DMP deleted
        401:
          $ref: "#/components/responses/AuthenticationRequiredResponse"
        403:
          $ref: "#/components/responses/InsufficientPermissionsResponse"
        404:
          $ref: "#/components/responses/DMPNotFoundResponse"
        406:
          $ref: "#/components/responses/NotAcceptableResponse"
        409:
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/GenericErrorResponse"
components:
  responses:
    DMPListResponse:
      description: "A list of DMPs"
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/DMPListResponseBody"
        application/json:
          schema:
            $ref: "#/components/schemas/DMPListResponseBody"
    AuthenticationRequiredResponse:
      description: |
        Authentication required to delete the DMP. The implementation should always return this error code for
        unauthenticated users as deletion always requires authentication.
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/AuthenticationRequiredError"
        application/json:
          schema:
            $ref: "#/components/schemas/AuthenticationRequiredError"
    BadRequestResponse:
      description: Bad request
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/BadRequestError"
        application/json:
          schema:
            $ref: "#/components/schemas/BadRequestError"
    ConflictResponse:
      description: |
        A precondition specified in the If-Unmodified-Since or If-Match header cannot be
        fulfilled. The client should apply the modifications again and resubmit the request.
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/ConflictError"
        application/json:
          schema:
            $ref: "#/components/schemas/ConflictError"
    DMPNotFoundResponse:
      description: |
        DMP not found or not enough permissions to delete the specified DMP.
        Security note: implementations may return this error code even if the DMP exists but the authenticated user
        has no permissions to view it in order to avoid leaking information about the existence of a DMP. This may
        especially be important if the DMP tool uses numeric IDs to avoid enabling an enumeration of possible DMP
        IDs.
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/DMPNotFoundError"
        application/json:
          schema:
            $ref: "#/components/schemas/DMPNotFoundError"
    GenericErrorResponse:
      description: The request failed
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/GenericError"
        application/json:
          schema:
            $ref: "#/components/schemas/GenericError"
    InsufficientPermissionsResponse:
      description: |
        Insufficient permissions to delete the specified a DMP.
        Note: the implementation may opt to return a 404 instead to avoid leaking information about the
        existence of a DMP with the specified ID if the user has no permissions to at least view this DMP. The
        implementation SHOULD return this error code if the user has at least read permissions for this DMP.
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/InsufficientPermissionsError"
        application/json:
          schema:
            $ref: "#/components/schemas/InsufficientPermissionsError"
    NotAcceptableResponse:
      description: |
        The server cannot produce a response that the client requested in its `Accept` header.
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/NotAcceptableError"
        application/json:
          schema:
            $ref: "#/components/schemas/NotAcceptableError"
    UnsupportedMediaTypeResponse:
      description: |
        The server cannot process the request the client because it does not support the DMP standard the client
        sent the document in.
      content:
        application/vnd.org.rd-alliance.dmp-common.v1.2+json:
          schema:
            $ref: "#/components/schemas/UnsupportedMediaTypeError"
        application/json:
          schema:
            $ref: "#/components/schemas/UnsupportedMediaTypeError"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DMPSystemID:
      type: string

    #endregion
    #region Subordinate objects
    DMPDocument:
      type: object
      properties:
        dmp:
          $ref: "#/components/schemas/DMPData"
    DMPWithID:
      allOf:
        - $ref: "#/components/schemas/DMPDocument"
        - type: object
          required:
            - id
          properties:
            id:
              title: "ID"
              description: "A system-dependent identifier for this DMP. The identifier is not valid outside of the current API, use the dmpId field in the DMP object itself it you need a system-independent identifier."
              $ref: "#/components/schemas/DMPSystemID"

    #endregion
    #region Non-error responses
    DMPListResponseBody:
      type: object
      required:
        - total_count
        - items
      properties:
        total_count:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/DMPWithID"

    #endregion
    #region Error responses
...
    AuthenticationRequiredError:
      title: "Authentication required"
      description: Authentication required to perform the specified request.
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required:
            - error_code
          properties:
            error_code:
              title: Error code
              type: string
              const: "authentication_required"
...
    NotAcceptableError:
      title: "Cannot meet Accept requirements"
      description: The client has requested a DMP standard that the server cannot fulfill.
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required:
            - error_code
          properties:
            error_code:
              title: Error code
              type: string
              const: "not_acceptable"
...
    BadRequestError:
      title: "Bad request"
      description: "The passed HTTP request was invalid."
      oneOf:
        - $ref: "#/components/schemas/InvalidQueryStringError"
        - $ref: "#/components/schemas/DMPInvalidError"
        - $ref: "#/components/schemas/GenericBadRequestError"
      discriminator:
        propertyName: error_code
        mapping:
          invalid_query_string: "#/components/schemas/InvalidQueryStringError"
          invalid_dmp: "#/components/schemas/DMPInvalidError"
          bad_request: "#/components/schemas/GenericBadRequestError"
...
    DMPInvalidError:
      title: "Invalid DMP"
      description: "The passed DMP was invalid"
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required:
            - error_code
          properties:
            error_code:
              title: Error code
              type: string
              const: "dmp_invalid"
...
    ContactID:
      type: object
      title: Contact ID
      properties:
        identifier:
          type: string
          title: Identifier
          examples:
            - 0000-0003-0644-4174
        type:
          type: string
          title: Type
          description: "Identifier type. Suggested values: orcid, isni, openid"
          examples:
            - orcid
      required:
        - identifier
        - type
...
    ContributorID:
      type: object
      title: Contributor ID
      properties:
        identifier:
          type: string
          title: Identifier
          examples:
            - 0000-0003-0644-4174
        type:
          type: string
          title: Type
          description: "Identifier type. Suggested values: orcid, isni, openid"
          examples:
            - orcid
      required:
        - identifier
        - type
...
    DatasetID:
      type: object
      title: Dataset ID
      properties:
        identifier:
          type: string
          title: Identifier
          examples:
            - 11353/10.923628
        type:
          type: string
          title: Type
          description: "Dataset identifier type. Suggested values: handle, doi, ark, url"
          examples:
            - handle
      required:
        - identifier
        - type
...
    DMPID:
      type: object
      title: DMP ID
      description: Identifier for the DMP itself
      properties:
        identifier:
          type: string
          title: Identifier
          examples:
            - 10.1371/journal.pcbi.1006750
        type:
          type: string
          title: Type
          description: "The DMP Identifier Type. Suggested values: handle, doi, ark, url"
          examples:
            - doi
      required:
        - identifier
        - type
...
    FunderID:
      type: object
      title: Funder ID
      description: Funder ID of the associated project
      properties:
        identifier:
          type: string
          title: Identifier
          description: "Funder ID, recommended to use CrossRef Funder Registry. See:
            https://www.crossref.org/services/funder-registry/"
          examples:
            - "501100002428"
        type:
          type: string
          title: Type
          description: "Identifier type. Suggested values: fundref, url"
          examples:
            - fundref
      required:
        - identifier
        - type
...
    GrantID:
      type: object
      title: Grant ID
      description: Grant ID of the associated project
      properties:
        identifier:
          type: string
          title: Identifier
          examples:
            - http://example.org/grants/776242
        type:
          type: string
          title: Type
          description: "Identifier type. Suggested values: url"
          examples:
            - url
      required:
        - identifier
        - type
...
    LanguageCode:
      type: string
      enum:
        - aar
        - abk
        - afr
        - aka
        - amh
        - ara
        - arg
        - asm
        - ava
        - ave
        - aym
        - aze
        - bak
        - bam
        - bel
        - ben
        - bih
        - bis
        - bod
        - bos
        - bre
        - bul
        - cat
        - ces
        - cha
        - che
        - chu
        - chv
        - cor
        - cos
        - cre
        - cym
        - dan
        - deu
        - div
        - dzo
        - ell
        - eng
        - epo
        - est
        - eus
        - ewe
        - fao
        - fas
        - fij
        - fin
        - fra
        - fry
        - ful
        - gla
        - gle
        - glg
        - glv
        - grn
        - guj
        - hat
        - hau
        - hbs
        - heb
        - her
        - hin
        - hmo
        - hrv
        - hun
        - hye
        - ibo
        - ido
        - iii
        - iku
        - ile
        - ina
        - ind
        - ipk
        - isl
        - ita
        - jav
        - jpn
        - kal
        - kan
        - kas
        - kat
        - kau
        - kaz
        - khm
        - kik
        - kin
        - kir
        - kom
        - kon
        - kor
        - kua
        - kur
        - lao
        - lat
        - lav
        - lim
        - lin
        - lit
        - ltz
        - lub
        - lug
        - mah
        - mal
        - mar
        - mkd
        - mlg
        - mlt
        - mon
        - mri
        - msa
        - mya
        - nau
        - nav
        - nbl
        - nde
        - ndo
        - nep
        - nld
        - nno
        - nob
        - nor
        - nya
        - oci
        - oji
        - ori
        - orm
        - oss
        - pan
        - pli
        - pol
        - por
        - pus
        - que
        - roh
        - ron
        - run
        - rus
        - sag
        - san
        - sin
        - slk
        - slv
        - sme
        - smo
        - sna
        - snd
        - som
        - sot
        - spa
        - sqi
        - srd
        - srp
        - ssw
        - sun
        - swa
        - swe
        - tah
        - tam
        - tat
        - tel
        - tgk
        - tgl
        - tha
        - tir
        - ton
        - tsn
        - tso
        - tuk
        - tur
        - twi
        - uig
        - ukr
        - urd
        - uzb
        - ven
        - vie
        - vol
        - wln
        - wol
        - xho
        - yid
        - yor
        - zha
        - zho
        - zul
      title: Language
      description: Language of the metadata expressed using ISO 639-3.
      examples:
        - eng
...
    MetadataStandardID:
      type: object
      title: Metadata Standard ID
      properties:
        identifier:
          type: string
          title: Identifier
          examples:
            - http://www.dublincore.org/specifications/dublin-core/dcmi-terms/
        type:
          type: string
          enum:
            - url
            - other
          title: Type
          description: "Identifier type. Suggested values: url"
          examples:
            - url
      required:
        - identifier
        - type
...